complement {munsell}	R Documentation

Find the complement of a munsell colour

Description

Finds the munsell colour with the same chroma and value but on the opposite side of the hue circle.

Usage

  complement(col, ...)

Arguments

col
character vector of Munsell colours

...
passed on to in_gamut. Use fix = TRUE to fix "bad" complement

Value

character vector of Munsell colours

Examples

complement("5PB 2/4")
cols <- c("5PB 2/4", "5Y 7/8")
plot_mnsl(c(cols, complement(cols)))

[Package munsell version 0.4 Index]