fix_mnsl {munsell}	R Documentation

Fix an undefined Munsell colour

Description

Takes correctly specified but undefined colours and outputs something sensible. Normally this happens when the chroma is too high. So, here sensible means the colour with the same hue and value and maximum defined chroma.

Usage

fix_mnsl(col)

Arguments

col
a character vector representing Munsell colours.

Value

a character vector containing the fixed colours.

Examples

fix_mnsl(c("5R 5/8","2.5R 9/28"))

[Package munsell version 0.5.0 Index]