hvc2mnsl {munsell}	R Documentation

Converts a hue, chroma and value to a Munsell colour

Description

Takes separate specifications of hue, value and chroma and returns the text specification of that colour.

Usage

hvc2mnsl(hue, value = NULL, chroma = NULL, ...)

Arguments

hue
a character vector of Munsell hues, or a 3 column data frame containing the hue value and chroma levels

value
a numeric vector of values

chroma
a numeric vector of chromas

...
passed on to check_mnsl. Use fix = TRUE to fix "bad" colours

Details

Munsell colours are specified by hue, value and chroma. They take a form like "5PB 5/10" where the first characters represent the hue, followed by a space then the value and chroma separated by a "/". In this package value should be an integer in 0:10 and chroma an even number at most 24. Note that not all possible specifications result in representable colours. Regular recycling rules apply.

Value

a character string specification of a hex colour

See Also

check_mnsl, mnsl2hex

Examples

hvc2mnsl("5PB", 5, 10)
# All values of 5PB with chroma 10
hvc2mnsl("5PB", 1:9, 10) # note some are undefined
plot_mnsl(hvc2mnsl("5PB", 1:9, 10))

[Package munsell version 0.5.0 Index]