in_gamut {munsell}	R Documentation

Checks if a Munsell colour is defined in RGB space

Description

Not all possible correctly formatted Munsell colours result in a colour representable in RGB space. This function checks if the colour is representable.

Usage

in_gamut(col, fix = FALSE)

Arguments

col
a character vector representing Munsell colours.

fix
passed on to fix_mnsl. Use fix = TRUE to fix "bad" colours

Value

a character vector containing the input colours. If any colours were outside the gamut they will be represented by NA.

Examples

in_gamut(c("5R 5/8","2.5R 9/28"))

[Package munsell version 0.5.0 Index]