munsell.map {munsell}	R Documentation

Provides link between Munsell colours hex and LUV

Description

Lookup table for converting Munsell colours to hex and LUV.

Usage

munsell.map

Format

A dataframe containing columns hue, chroma, value, hex, name, L, U, V for 2740 Munsell colours. Conversion from Munsell Renotation Data is done using formulae at http://www.brucelindbloom.com/index.html?ColorCalcHelp.html and the function hex in the package colorspace. Code for conversion can be found in the package directory in inst/raw/getmunsellmap.r

Source

Munsell Renotation Data, RIT Munsell Color Science Laboratory. http://www.cis.rit.edu/mcsl/online/munsell.php

[Package munsell version 0.4 Index]