chroma_slice {munsell}	R Documentation

Plot all colours with the same chroma

Description

Plots slices of the Munsell colour system where chroma is constant.

Usage

chroma_slice(chroma.name = seq(0, 26, by = 2), back.col = "white")

Arguments

chroma.name
integer vector of the desired values.

back.col
colour for the background

Value

ggplot object

Examples

chroma_slice(2)
chroma_slice(18)
# Maybe want to delete text and add axis instead
p <- chroma_slice(18)
p$layers[[2]] <- NULL # remove text layer
p + ggplot2::theme(axis.text = ggplot2::element_text(), 
   axis.text.x = ggplot2::element_text(angle = 90, hjust = 1))
# all values 
## Not run: chroma_slice(seq(0, 38, by = 2))

[Package munsell version 0.5.0 Index]