mnsl2hvc {munsell}	R Documentation

Converts a Munsell colour to a hue, chroma and value triplet

Description

Takes a text specification of a Munsell colour and returns the hue, chroma and value triplet.

Usage

mnsl2hvc(col, ...)

Arguments

col
a character vector of Munsell colours

...
passed on to check_mnsl. Use fix = TRUE to fix "bad" colours

Details

Munsell colours are specified by hue, value and chroma. They take a form like "5PB 5/10" where the first characters represent the hue, followed by a space then the value and chroma separated by a "/". In this package value should be an integer in 0:10 and chroma an even number at most 24. Note that not all possible specifications result in representable colours.

Value

a data frame with named columns hue, value and chroma containing the hue, value and chroma levels.

See Also

check_mnsl, hvc2mnsl

Examples

mnsl2hvc("5PB 5/10")
hvc2mnsl(mnsl2hvc("5PB 5/10"))

[Package munsell version 0.5.0 Index]