plot_closest {munsell}	R Documentation

Plot closest Munsell colour to an sRGB colour

Description

Take an sRGB colour and plots it along with the closest Munsell colour (using rgb2mnsl to find it)

Usage

plot_closest(R, G = NULL, B = NULL, back.col = "white")

Arguments

R
a numeric vector of red values or a 3 column matrix with the proportions R, G, B in the columns.

G
numeric vector of green values

B
numeric vector of blue values

back.col
colour for the background

Value

ggplot object

See Also

rgb2mnsl

Examples

plot_closest(0.1, 0.1, 0.3)
plot_closest(matrix(c(.1, .2, .4, .5, .6, .8),  ncol = 3)) 

[Package munsell version 0.5.0 Index]