plot_mnsl {munsell}	R Documentation

Plot a munsell colour

Description

Takes munsell text specifications and plots colour squares of them.

Usage

plot_mnsl(cols, back.col = "white", ...)

Arguments

cols
character vector specifying colours in Munsell form

back.col
specification of background colour of display

...
passed to check_mnsl. Add fix = TRUE to fix "bad" colours()

Value

A ggplot object

Examples

plot_mnsl("5R 5/6")
plot_mnsl("5R 5/6",  back.col = "grey40")
p <- plot_mnsl(c("5R 6/6", "5Y 6/6", "5G 6/6", "5B 6/6", "5P 6/6"),
 back.col = "grey40")
p
# returned object is a ggplot object so we can alter the layout
summary(p)
p + ggplot2::facet_wrap(~ num, nrow = 1)

[Package munsell version 0.5.0 Index]