rgb2mnsl {munsell}	R Documentation

Converts an sRGB colour to Munsell

Description

Finds the closest Munsell colour (in LUV space) to the specified sRGB colour

Usage

rgb2mnsl(R, G = NULL, B = NULL)

Arguments

R
a numeric vector of red values or a 3 column matrix with the proportions R, G, B in the columns.

G
numeric vector of green values

B
numeric vector of blue values

See Also

plot_closest

Examples

rgb2mnsl(0.1, 0.1, 0.3)
rgb2mnsl(matrix(c(.1, .2, .4, .5, .6, .8),  ncol = 3))
plot_closest(matrix(c(.1, .2, .4, .5, .6, .8),  ncol = 3))

[Package munsell version 0.5.0 Index]