seq_mnsl {munsell}	R Documentation

Generate a sequence of Munsell colours

Description

Generates a sequence of Munsell colours. The sequence is generated by finding the closest munsell colours to a equidistant sequence of colours in #' LUV space.

Usage

seq_mnsl(from, to, n, fix = FALSE)

Arguments

from
character string of first Munsell colour

to
character string of last Munsell colour

n
number of colours in sequence

fix
Should colours outside of the gamut be fixed? Passed on to fix_mnsl

Value

character vector of Munsell colours

Examples

seq_mnsl("5R 2/4", "5R 5/16", 4)
plot_mnsl(seq_mnsl("5R 2/4", "5R 5/16", 4))
plot_mnsl(seq_mnsl("5R 5/6", 
  complement("5R 5/6"), 5))

[Package munsell version 0.5.0 Index]